pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "CyLinter"
version = "0.0.47"
description = "CyLinter: An Interactive Image Segmentation Filter for Multiplex Microscopy"
readme = "README.md"
license = "MIT"
keywords =['CyLinter multiplex microscopy quality control']
classifiers=[
    'Development Status :: 4 - Beta',
    'Intended Audience :: End Users/Desktop',
    'Intended Audience :: Science/Research',
    'Framework :: napari',
    'License :: OSI Approved :: MIT License',
    'Natural Language :: English',
    'Operating System :: OS Independent',
    'Programming Language :: Python :: 3',
    'Topic :: Scientific/Engineering :: Visualization'
]
authors = ["Gregory J. Baker <gregory_baker2@hms.harvard.edu>"]
homepage = "https://github.com/labsyspharm/cylinter"

[tool.poetry.dependencies]
cellcutter = "*"
hdbscan = "*"
joblib = "*"
magicgui = "*"
matplotlib = "<3.6"
napari = { version = "*", extras = ["all"] }
numpy = "<1.24"
natsort = "*"
numba = "*"
pandas = "*"
pyqt = "*"
pyarrow = "*"
pyyaml = "*"
qtpy = "*"
scikit-image = "*"
scikit-learn = "<1.3"
scipy = "*"
seaborn = "*"
tifffile = "*"
umap-learn = "*"
zarr = "<2.15"
pydantic = "<2.0"
bokeh = "*"
svglib = "*"
pypdf2 = "*"

[tool.poetry.scripts]
cylinter = "cylinter.cylinter:main"
